tkinter notepad/requirements.txt
altgraph==0.17.2
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
Django==4.0.1
docopt==0.6.2
domain==0.1.4
EasyProcess==0.3
entrypoint2==0.2.4
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.28.3
future==0.18.2
idna==3.3
ip-address==1.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mss==6.1.0
mysql-connector-python==8.0.27
numpy==1.21.4
packaging==21.3
pandas==1.3.5
pefile==2021.9.3
Pillow==8.4.0
pipenv==2022.1.8
pipreqs==0.4.11
platformdirs==2.4.1
protobuf==3.19.1
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pynput==1.7.6
pyparsing==3.0.6
pyscreenshot==3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-futures==1.0.0
scikit-learn==1.0.2
scipy==1.7.3
selenium==3.141.0
six==1.16.0
sklearn==0.0
sounddevice==0.4.4
sqlparse==0.4.2
stem==1.8.0
threadpoolctl==3.0.0
tld==0.12.6
torrequest==0.1.0
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.13.0
virtualenv-clone==0.5.7
website==1.1
Werkzeug==2.0.2
whois==0.9.13
yarg==0.1.9
